fix: move circular dependency import to own module
