+++ apps/vf-features/__tests__/working.profile.page.test.tsx
@@ -1,6 +1,6 @@
 import WorkingProfilePage from '@pages/profile/working-profile.page';
 import userEvent from '@testing-library/user-event';
-import * as UtilsExports from '@/lib/utils';
+import * as UtilsExports from '@/lib/utils/auth';
 import { EMPLOYMENT_TYPE_LABELS } from '@shared/lib/constants';
 import {
   MOCK_AUTH_STATE,
@@ -16,7 +16,7 @@ describe('Personal profile page', () => {
   it.only('renders a working profile page / form for authenticated user, with correct profile values', async () => {
     jest
       .spyOn(UtilsExports, 'getValidAuthState')
-      .mockImplementation(() => MOCK_AUTH_STATE);
+      .mockImplementation(async () => MOCK_AUTH_STATE);
 
     const user = userEvent.setup();
