+++ apps/vf-mvp/__tests__/personal.profile.page.test.tsx
@@ -1,5 +1,5 @@
 import PersonalProfilePage from '@mvp/pages/profile/personal-profile.page';
-import * as UtilsExports from '@/lib/utils';
+import * as UtilsExports from '@/lib/utils/auth';
 import {
   MOCK_AUTH_STATE,
   MOCK_ISO_COUNTRIES,
@@ -14,7 +14,7 @@ describe('Personal profile page', () => {
   it('renders a personal profile page / form for authenticated user, with correct profile values', async () => {
     jest
       .spyOn(UtilsExports, 'getValidAuthState')
-      .mockImplementation(() => MOCK_AUTH_STATE);
+      .mockImplementation(async () => MOCK_AUTH_STATE);
 
     renderWithProviders(<PersonalProfilePage />);
