+++ apps/vf-mvp/__tests__/working.profile.page.test.tsx
@@ -1,6 +1,6 @@
 import WorkingProfilePage from '@mvp/pages/profile/working-profile.page';
 import userEvent from '@testing-library/user-event';
-import * as UtilsExports from '@/lib/utils';
+import * as UtilsExports from '@/lib/utils/auth';
 import { EMPLOYMENT_TYPE_LABELS } from '@shared/lib/constants';
 import {
   MOCK_AUTH_STATE,
